Up to the captain - Everything works at this point
Changed Hallway1 to work off an If statement. This checks if the player has killed the guard or not. Added a health potion after the guard fight. Heals player to full.

Added more to the hallway and a new variable to check for the key.

Everything Up until the captains quarters works at this point.