Door check, extra couts, pressEnter added
The player now doesn't know that the door is locked. After the door is checked the couts will call it a LOCKED door. Added couts to display something different if the player is coming from a different location. Added pressEnter function that allows user to press ENTER (or enter anything) and it will continue the program. Edited a few things as well.