Inn map Added
Added a map for the Inn.

Moved the Inn keeper bools to main and sent them through that way we can send them where needed.

Added comments with direction on how to end the loop/function. This is important.

Also side note, currently the town map breaks for the the builidings. Will look at fixes at a later time.