Polished Inn Function
Formatted the Inn function and added some endl statements to it to make it a bit neater. Also added a couple pressenter() to it to make it flow nicer.

Changed the console window size to start a bit bigger than it was before. This is to properly display all the Inn stuff